Дробит код на модули
